+++ Pipfile
@@ -31,9 +31,10 @@ coa-django-countable_field = "==1.6"
 django-flags = "*"
 pympler = "*"
 psycopg2 = "==2.8.4"
-wagtail-factories = "*"
+wagtail-factories = {editable = true,ref = "dev",git = "https://github.com/nhsuk/wagtail-factories.git"}
 scout-apm = "*"
 django-silk = "*"
+l18n = "*"
 
 [requires]
 python_version = "3.7"
